api to upload media for comment; refactor to /api/media endpoint
