+++ mini_gplus/resources/posts.py
@@ -1,8 +1,5 @@
-import os
-import boto3
 import werkzeug
 import uuid
-import json
 from flask_restful import reqparse, Resource, fields, marshal_with, marshal
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from mini_gplus.daos.user import find_user
@@ -12,90 +9,15 @@ from mini_gplus.daos.post import dangerously_get_post, create_post, sees_post, r
     retrieves_posts_on_profile, delete_post
 from mini_gplus.daos.post_cache import get_in_post_cache
 from mini_gplus.daos.circle_cache import get_in_circle_cache
-from mini_gplus.daos.media import get_media
 from mini_gplus.daos.exceptions import UnauthorizedAccess
-from mini_gplus.utils.now_ms import now_ms
-from mini_gplus.utils.profiling import timer
 from .users import user_fields
 from .s3 import upload_to_s3, delete_from_s3
 from .pagination import pagination_parser
 from .mention import check_mentioned_user_ids
 from .comments import comment_fields
-from .media_url_cache import r, RMediaUrl
+from .media import check_media_object_names, MediaUrls
 
-MaxPostMediaCount = 4
-PostMediaUrlExpireSeconds = 3600 * 12  # 12 hours
-
-
-# Cache structure within Redis
-# "mediaUrl" -> object_name -> "media url"(space)"media url generated time in ms"
-
-
-class MediaUrls(fields.Raw):
-    def format(self, media_list):
-        if not media_list:
-            return []
-
-        @timer
-        def get_media_url(media):
-            object_name = media.id
-            # subtract expiry by 10 seconds for some network overhead
-            r_media_url = r.hget(RMediaUrl, object_name)
-            if r_media_url:
-                r_media_url = r_media_url.decode('utf-8')
-                if now_ms() < int(r_media_url.split(" ")[1]) + (PostMediaUrlExpireSeconds - 10) * 1000:
-                    return r_media_url.split(" ")[0]
-
-            sts_client = boto3.client(
-                'sts',
-                endpoint_url=os.environ['STS_ENDPOINT_URL'],
-                region_name=os.environ.get('AWS_REGION', ''),
-                aws_access_key_id=os.environ['AWS_ACCESS_KEY'],
-                aws_secret_access_key=os.environ['AWS_SECRET_KEY']
-            )
-            s3_bucket_name = os.environ['S3_BUCKET_NAME']
-
-            # obtain temp token
-            read_media_policy = {
-                "Version": "2012-10-17",
-                "Statement": [
-                    {
-                        "Effect": "Allow",
-                        "Action": "s3:GetObject",
-                        "Resource": [f"arn:aws:s3:::{s3_bucket_name}/{object_name}"],
-                    },
-                ],
-            }
-            assume_role_response = sts_client.assume_role(
-                # for minio this is moot
-                # for s3 this role allows all media read, but intersects with the inline policy, the temp role
-                #    would still be minimal privilege
-                RoleArn=os.environ['MEDIA_READER_ROLE_ARN'],
-                # media-reader is the only principal who can assume the role so this can be fixed
-                RoleSessionName='media-reader',
-                Policy=json.dumps(read_media_policy),
-                DurationSeconds=PostMediaUrlExpireSeconds,
-            )
-            temp_s3_client = boto3.client(
-                's3',
-                endpoint_url=os.environ['S3_ENDPOINT_URL'],
-                region_name=os.environ.get('AWS_REGION', ''),
-                aws_access_key_id=assume_role_response['Credentials']['AccessKeyId'],
-                aws_secret_access_key=assume_role_response['Credentials']['SecretAccessKey'],
-                aws_session_token=assume_role_response['Credentials']['SessionToken'],
-            )
-
-            # get pre-signed url
-            media_url = temp_s3_client.generate_presigned_url(
-                ClientMethod='get_object',
-                Params={'Bucket': s3_bucket_name, 'Key': media.id},
-                ExpiresIn=PostMediaUrlExpireSeconds
-            )
-
-            r.hset(RMediaUrl, object_name, f"{media_url} {now_ms()}")
-            return media_url
-
-        return list(map(get_media_url, media_list))
+MaxMediaCount = 4
 
 
 class ResharedFrom(fields.Raw):
@@ -194,12 +116,6 @@ class Posts(Resource):
         media_object_names = args['media_object_names']
         if reshared_from and media_object_names:
             return {'msg': "Reshared post is not allowed to have media"}, 400
-        media_objects = []
-        for media_object_name in media_object_names:
-            media_object = get_media(media_object_name)
-            if not media_object:
-                return {"msg": f"Media {media_object_name} is not found"}, 404
-            media_objects.append(media_object)
 
         post = create_post(
             user,
@@ -208,7 +124,7 @@ class Posts(Resource):
             circles=circles,
             reshareable=args['reshareable'],
             reshared_from=reshared_from_post,
-            media_list=media_objects,
+            media_list=check_media_object_names(media_object_names),
             mentioned_users=check_mentioned_user_ids(args['mentioned_user_ids'])
         )
         if not post:
@@ -216,18 +132,18 @@ class Posts(Resource):
         return post, 201
 
 
-post_media_parser = reqparse.RequestParser()
-for i in range(MaxPostMediaCount):
-    post_media_parser.add_argument('media' + str(i), type=werkzeug.datastructures.FileStorage, location='files',
-                                   required=False, default=None)
+media_parser = reqparse.RequestParser()
+for i in range(MaxMediaCount):
+    media_parser.add_argument('media' + str(i), type=werkzeug.datastructures.FileStorage, location='files',
+                              required=False, default=None)
 
 
-class PostMedia(Resource):
+class Media(Resource):
     @jwt_required()
     def post(self):
-        args = post_media_parser.parse_args()
+        args = media_parser.parse_args()
         media_files = []
-        for i in range(MaxPostMediaCount):
+        for i in range(MaxMediaCount):
             media_file = args['media' + str(i)]
             if media_file:
                 media_files.append(media_file)
@@ -278,7 +194,6 @@ class Post(Resource):
         post = dangerously_get_post(post_id)
         if user != post.author:
             raise UnauthorizedAccess()
-
         for m in post.media_list:
             delete_from_s3(m)
